Refactor: 로그인, 로그아웃 관련 securityConfig 수정
